backend/requirements.txt
fastapi==0.129.2
uvicorn==0.41.0
agno==2.5.3
groq==1.0.0
python-dotenv==1.2.1
pandas==3.0.1
simple-icd-10==2.1.1
python-multipart==0.0.22
pydantic==2.12.5
pydantic-settings==2.13.1
